hydi-voice/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
openai>=1.30.0
httpx>=0.27.0
qrcode==7.4.2
pillow==10.4.0
bleak==0.22.2